+++ CHANGES.rst
@@ -33,6 +33,7 @@ Added
 * Added CHOOSE() function
 * Added FORECAST() function
 * Added INTERCEPT() function
+* Added IFNA() function
 * Added ISBLANK() function
 * Added ISLOGICAL() function
 * Added ISNONTEXT() function
@@ -42,19 +43,24 @@ Added
 * Added SUBSTITUTE() function
 * Added TEXT() function  (Thanks, Luckykarter)
 * Added TREND() function
+* Added Reference Form for INDEX()
 * Added str_params to excel_helper()
+* Added ExcelCompiler.validate_serialized()
 
 Changed
 -------
 
 * Improve LINEST() compatibilty w/ Excel
 * Improve error and number handling in some Text functions
+* Improve IFS() to support array context
+* Missing references from INDIRECT() and OFFSET() resolve more often
 
 Fixed
 -----
 
 * Fix #111, Incorrect implementation of YEARFRAC
 * Fixed some exceptions in LINEST()
+* Fix serialize ranges with formulas
 
 
 [1.0b26] - 2021-06-18